+++ apps/taxes/src/modules/taxjar/taxjar-order-created-adapter.test.ts
@@ -1,169 +1,3 @@
-import { describe, it, expect } from "vitest";
-import {
-  sumPayloadLines,
-  TaxJarOrderCreatedPayloadTransformer,
-} from "./taxjar-order-created-adapter";
-import { OrderStatus } from "../../../generated/graphql";
-
-const MOCKED_PAYLOAD = {
-  order: {
-    id: "T3JkZXI6OTU4MDA5YjQtNDUxZC00NmQ1LThhMWUtMTRkMWRmYjFhNzI5",
-    created: "2023-04-11T11:03:09.304109+00:00",
-    status: OrderStatus.Unfulfilled,
-    user: {
-      id: "VXNlcjo5ZjY3ZjY0Zi1iZjY5LTQ5ZjYtYjQ4Zi1iZjY3ZjY0ZjY0ZjY=",
-      email: "tester@saleor.io",
-    },
-    channel: {
-      id: "Q2hhbm5lbDoy",
-      slug: "channel-pln",
-    },
-    shippingAddress: {
-      streetAddress1: "123 Palm Grove Ln",
-      streetAddress2: "",
-      city: "LOS ANGELES",
-      countryArea: "CA",
-      postalCode: "90002",
-      country: {
-        code: "US",
-      },
-    },
-    billingAddress: {
-      streetAddress1: "123 Palm Grove Ln",
-      streetAddress2: "",
-      city: "LOS ANGELES",
-      countryArea: "CA",
-      postalCode: "90002",
-      country: {
-        code: "US",
-      },
-    },
-    total: {
-      net: {
-        amount: 183.33,
-      },
-      tax: {
-        amount: 12.83,
-      },
-      currency: "USD",
-    },
-    shippingPrice: {
-      gross: {
-        amount: 48.33,
-      },
-      net: {
-        amount: 43.74,
-      },
-    },
-    lines: [
-      {
-        productSku: "328223581",
-        productName: "Monospace Tee",
-        quantity: 1,
-        unitPrice: {
-          net: {
-            amount: 90,
-          },
-        },
-        totalPrice: {
-          net: {
-            amount: 90,
-          },
-          tax: {
-            amount: 8.55,
-          },
-        },
-      },
-      {
-        productSku: "328223580",
-        productName: "Monospace Tee",
-        quantity: 1,
-        unitPrice: {
-          net: {
-            amount: 45,
-          },
-        },
-        totalPrice: {
-          net: {
-            amount: 45,
-          },
-          tax: {
-            amount: 4.28,
-          },
-        },
-      },
-    ],
-    discounts: [
-      {
-        amount: {
-          amount: 10,
-        },
-        id: "RGlzY291bnREaXNjb3VudDox",
-      },
-    ],
-  },
-  channelConfig: {
-    providerInstanceId: "b8c29f49-7cae-4762-8458-e9a27eb83081",
-    enabled: false,
-    address: {
-      country: "US",
-      zip: "92093",
-      state: "CA",
-      city: "La Jolla",
-      street: "9500 Gilman Drive",
-    },
-  },
-};
-
-describe("TaxJarOrderCreatedPayloadTransformer", () => {
-  it("returns the correct order amount", () => {
-    const transformer = new TaxJarOrderCreatedPayloadTransformer();
-    const transformedPayload = transformer.transform(MOCKED_PAYLOAD);
-
-    expect(transformedPayload).toBe(183.33);
-  });
-});
-
-describe("sumPayloadLines", () => {
-  it("returns the sum of all line items when items quantity = 1", () => {
-    const result = sumPayloadLines([
-      {
-        quantity: 1,
-        unit_price: 90.45,
-        product_identifier: "328223581",
-      },
-      {
-        quantity: 1,
-        unit_price: 45.25,
-        product_identifier: "328223580",
-      },
-    ]);
-
-    expect(result).toBe(135.7);
-  });
-  it("returns the sum of all line items when items quantity > 1", () => {
-    const result = sumPayloadLines([
-      {
-        quantity: 3,
-        unit_price: 90.45,
-        product_identifier: "328223581",
-      },
-      {
-        quantity: 2,
-        unit_price: 45.25,
-        product_identifier: "328223580",
-      },
-      {
-        quantity: 1,
-        unit_price: 50.25,
-        product_identifier: "328223580",
-      },
-    ]);
-
-    expect(result).toBe(412.1);
-  });
-});
-
-describe.todo("TaxJarOrderCreatedResponseTransformer", () => {});
+import { describe } from "vitest";
 
 describe.todo("TaxJarOrderCreatedAdapter", () => {});
