+++ apps/taxes/src/modules/taxjar/taxjar-order-created-response-transformer.ts
@@ -1,0 +1,10 @@
+import { CreateOrderRes } from "taxjar/dist/types/returnTypes";
+import { CreateOrderResponse } from "../taxes/tax-provider-webhook";
+
+export class TaxJarOrderCreatedResponseTransformer {
+  transform(response: CreateOrderRes): CreateOrderResponse {
+    return {
+      id: response.order.transaction_id,
+    };
+  }
+}
